pyproject.toml
[project]
name = "computer-vision"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12,<3.14"
authors = [{ name = "Tim Boleslawsky", email = "tim.boleslawsky@yahoo.com" }]
dependencies = [
    "pytorch-lightning>=2.6.0",
    "pyyaml>=6.0.3",
    "tensorboard>=2.20.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
]
